+++ test/unit/lib/models/User.mocks.js
@@ -15,6 +15,7 @@ const Mock = function () {
   UserStub.find = sandbox.stub().usingPromise().resolves()
   UserStub.findById = sandbox.stub().usingPromise().resolves()
   UserStub.findOneAndRemove = sandbox.stub().usingPromise().resolves()
+  UserStub.findOneAndUpdate = sandbox.stub().usingPromise().resolves()
 
   const stubs = {
     user: userStub,